Checking for the regexps limits on every pcre call
As reported by Bruno de Almeida SecPcreMatchLimit was not working as expected,
however, the limits placed during the compilation time was working like a
charm. Looking at the code it seems that the match limit were placed to just
few calls of the libpcre. This patch make 100% of the calls to respect
the limits imposed by SecPcreMatchLimit and SecPcreMatchLimitRecursion.